+++ lib/types/index.dart
@@ -1,2 +1,3 @@
 export './models/index.dart';
 export './enums/index.dart';
+export './extensions/index.dart';